% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate.R
\name{create_plate}
\alias{create_plate}
\title{Create a MycoHalo plate object}
\usage{
create_plate(image)
}
\arguments{
\item{image}{A magick-image object.}
}
\value{
A mycohalo_plate object.
}
\description{
Create a MycoHalo plate object
}
