% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{delta_e76}
\alias{delta_e76}
\title{Colour difference CIE76 (Euclidean distance in L\emph{a}b*)}
\usage{
delta_e76(lab1, lab2)
}
\arguments{
\item{lab1, lab2}{n x 3 matrices (or vectors of length 3, recycled).}
}
\value{
Numeric vector of \eqn{\Delta E^*_{ab}}; ~2.3 is a just-noticeable
difference (Sharma 2003).
}
\description{
Colour difference CIE76 (Euclidean distance in L\emph{a}b*)
}
\examples{
delta_e76(c(50, 0, 0), c(52, 1, -1))
}
