% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_info.R
\name{image_info}
\alias{image_info}
\title{Extract image metadata}
\usage{
image_info(image)
}
\arguments{
\item{image}{A magick-image object.}
}
\value{
A tibble containing image metadata.
}
\description{
Retrieve metadata from a plate image.
}
\examples{
\dontrun{

img <- read_plate(
  "inst/extdata/control_01.jpg"
)

image_info(img)

}

}
