% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate.R
\name{is_mycohalo_plate}
\alias{is_mycohalo_plate}
\title{Test for a MycoHalo plate object}
\usage{
is_mycohalo_plate(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{mycohalo_plate}.
}
\description{
Test for a MycoHalo plate object
}
\examples{
is_mycohalo_plate(1)
}
