% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{lab_to_srgb}
\alias{lab_to_srgb}
\title{Convert CIE L\emph{a}b* to sRGB}
\usage{
lab_to_srgb(lab)
}
\arguments{
\item{lab}{Numeric n x 3 matrix (L, a, b) or H x W x 3 array.}
}
\value{
sRGB values in [0, 1], same shape as input.
}
\description{
Inverse of \code{\link[=srgb_to_lab]{srgb_to_lab()}}; out-of-gamut values are clipped to [0, 1].
}
\examples{
lab_to_srgb(cbind(L = 50, a = 0, b = 0))
}
