% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mycohalo_plate.R
\name{print.mycohalo_plate}
\alias{print.mycohalo_plate}
\title{Print a MycoHalo plate}
\usage{
\method{print}{mycohalo_plate}(x, ...)
}
\arguments{
\item{x}{A mycohalo_plate object.}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Print a MycoHalo plate
}
