% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_grayscale.R
\name{to_grayscale}
\alias{to_grayscale}
\title{Convert image to grayscale}
\usage{
to_grayscale(plate)
}
\arguments{
\item{plate}{A mycohalo_plate object.}
}
\value{
Updated mycohalo_plate object.
}
\description{
Convert an EBImage image to grayscale.
}
