% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_plate.R
\name{detect_plate}
\alias{detect_plate}
\title{Detect Petri dish}
\usage{
detect_plate(plate)
}
\arguments{
\item{plate}{A mycohalo_plate object.}
}
\value{
Updated mycohalo_plate object.
}
\description{
Detect the Petri dish region from the image background.
}
\examples{
\dontrun{
plate <- read_plate("plate.jpg")
plate <- detect_plate(plate)
}

}
