% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plate.R
\name{example_plate}
\alias{example_plate}
\title{Load example plate image}
\usage{
example_plate(type = c("control", "confrontation"))
}
\arguments{
\item{type}{Character scalar.
Either \code{"control"} or \code{"confrontation"}.}
}
\value{
A \code{mycohalo_plate} object.
}
\description{
Load an example image distributed with MycoHalo.
}
\examples{
\dontrun{
plate <- example_plate("control")
}

}
