% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mycohalo_plate.R
\name{is_mycohalo_plate}
\alias{is_mycohalo_plate}
\title{Check whether an object is a MycoHalo plate}
\usage{
is_mycohalo_plate(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Check whether an object is a MycoHalo plate
}
\examples{
is_mycohalo_plate(mtcars)

}
