% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate.R
\name{read_plate}
\alias{read_plate}
\title{Read a plate image}
\usage{
read_plate(path)
}
\arguments{
\item{path}{Character scalar.
Path to an image file.}
}
\value{
A \code{mycohalo_plate} object.
}
\description{
Import a fungal colony plate image and create a
MycoHalo plate object.
}
\examples{
\dontrun{
plate <- read_plate(
  "inst/extdata/control/control_01.jpg"
)
}

}
